Выполнил домашнюю работу Модуляция. Создал несколько модулей и подключил их в main.js
